chore: rename ESLint config file and update plugins for React and TypeScript
